Flask apis

+++ app/views/user_views.py
@@ -1,8 +1,9 @@
-from flask import Blueprint, jsonify, request, flash
-from flask_login import login_user, logout_user, login_required, current_user
+from flask import Blueprint, jsonify, request
+from flask_login import login_user, logout_user,current_user
+from flask_jwt_extended import create_access_token,jwt_required,get_jwt_identity
 from ..db_services.user_service import get_all_users, create_user, get_user_by_id, update_user, delete_user, authenticate_user
 from ..route.tasks import user_blueprint
-from flask_jwt_extended import create_access_token
+from app.decorators import admin_required,manager_required,employee_required
 
 @user_blueprint.route('/login', methods=['POST'])
 def login():
@@ -16,20 +17,25 @@ def login():
         return jsonify(message='Login successful', access_token=access_token)
     else:
         return jsonify(error='Invalid email or password')
+    
 
 @user_blueprint.route('/logout')
-@login_required
+@jwt_required()
+@admin_required
 def logout():
     logout_user()
-    flash('Logout successful', 'success')
     return jsonify(message='Logout successful')
 
-@user_blueprint.route('/alls', methods=["GET"])
+
+@user_blueprint.route('/all', methods=["GET"])
+@jwt_required()
 def manage_users():
-    if request.method == "GET":
+        current_user_id = get_jwt_identity()
         users_list = get_all_users()
         return jsonify(users_list=users_list)
-@user_blueprint.route('/creates', methods=["POST"])
+
+
+@user_blueprint.route('/create', methods=["POST"])
 def create_users():
         try:
             data = request.get_json()
@@ -39,27 +45,39 @@ def create_users():
                 password=data['password'],  
                 firstname=data['firstname'],
                 lastname=data['lastname'],
-                mobileno=data['mobileno']
+                mobileno=data['mobileno'],
+                role=data['role']
             )
             return jsonify(result)
         except Exception as e:
             return jsonify(error=f"Error creating user: {e}")
+        
 
-@user_blueprint.route('id/<string:user_id>', methods=["GET", "PUT", "DELETE"])
+@user_blueprint.route('user/<string:user_id>', methods=["GET", "PUT", "DELETE"])
+@jwt_required()
+@employee_required
 def manage_user(user_id):
-    if request.method == "GET":
-        user = get_user_by_id(user_id)
-        if user:
-            return jsonify(user=user)
-        else:
-            return jsonify(error=f"No user found with id {user_id}.")
-    elif request.method == "PUT":
-        try:
-            data = request.get_json()
-            result = update_user(user_id, email=data.get('email'))
+    current_user_id = get_jwt_identity()
+    if current_user_id:
+        if request.method == "GET":
+
+            user = get_user_by_id(user_id)
+            if user:
+                return jsonify(user=user)
+            else:
+                return jsonify(error=f"No user found with id {user_id}.")
+        elif request.method == "PUT":
+            try:
+                data = request.get_json()
+                result = update_user(
+                    user_id, mobileno=data.get('mobileno'),
+                    username=data.get('username'),
+                    firstname=data.get('firstname'),
+                    lastname=data.get('lastname')
+                    )
+                return jsonify(result)
+            except Exception as e:
+                return jsonify(error=f"Error updating user: {e}")
+        elif request.method == "DELETE":
+            result = delete_user(user_id)
             return jsonify(result)
-        except Exception as e:
-            return jsonify(error=f"Error updating user: {e}")
-    elif request.method == "DELETE":
-        result = delete_user(user_id)
-        return jsonify(result)
